pyproject.toml
[project]
name = "youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ddgs>=9.10.0",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
